backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Django backend for Factory managemnt system."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "channels>=4.3.1",
    "channels-redis>=4.3.0",
    "daphne>=4.2.1",
    "dj-rest-auth>=6.0.0",
    "django>=5.2.6",
    "django-allauth==0.61.1",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.1.1",
    "whitenoise[brotli]>=6.10.0",
]

[dependency-groups]
dev = [
    "faker>=37.8.0",
    "rich>=14.2.0",
]
